+++ plone/i18n/normalizer/tests/test_normalizer.py
@@ -13,6 +13,12 @@ import six
 import unittest
 
 
+def utf8unicode(value):
+    if six.PY2:
+        return unicode(value, 'utf-8')
+    return value
+
+
 def testIDNormalizer():
     """
       >>> util = queryUtility(IIDNormalizer)
@@ -31,7 +37,7 @@ def testIDNormalizer():
       >>> util.normalize(u">>>here'!--s yet another!!!")
       'here-s-yet-another'
 
-      >>> util.normalize(six.text_type("umläut.doc", 'utf-8'))
+      >>> util.normalize(utf8unicode("umläut.doc"))
       'umlaut-doc'
 
       >>> from plone.i18n.normalizer import MAX_LENGTH
@@ -54,7 +60,7 @@ def testIDNormalizer():
       >>> len(util.normalize(testString)) <= MAX_LENGTH
       True
 
-      >>> util.normalize(six.text_type("rest `n` peace", 'utf-8'))
+      >>> util.normalize(u"rest `n` peace")
       'rest-n-peace'
 
       >>> util.normalize(u"short-hello-version", max_length=10)
@@ -82,13 +88,13 @@ def testLocaleAwareIDNormalizer():
       >>> util.normalize(u'simpleandsafe', locale='de')
       'simpleandsafe'
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de')
       'text-with-umlaeut'
 
     Make sure we get the de normalizer as there's no special one for de-DE
     registered.
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de-DE')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de-DE')
       'text-with-umlaeut'
 
       >>> util.normalize(u'simpleandsafe', locale='pt-BR')
@@ -106,7 +112,7 @@ def testLocaleAwareIDNormalizer():
     Make sure we also handle POSIX-format locale identifiers,
     for backwards-compatibility with an earlier version of plone.i18n.
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de_DE')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de_DE')
       'text-with-umlaeut'
 
     """
@@ -136,13 +142,13 @@ def testFileNameNormalizer():
       >>> util.normalize("pseudo_filename,pot,#1.doc")
       'pseudo_filename-pot-#1.doc'
 
-      >>> util.normalize(six.text_type("umläut.doc", 'utf-8'))
+      >>> util.normalize(utf8unicode("umläut.doc"))
       'umlaut.doc'
 
       >>> len(util.normalize(u'aa' * 2000))
       1023
 
-      >>> util.normalize(six.text_type("rest `n` peace", 'utf-8'))
+      >>> util.normalize(u"rest `n` peace")
       'rest -n- peace'
 
       >>> util.normalize(u"short-hello-version", max_length=10)
@@ -165,13 +171,13 @@ def testLocaleAwareFileNameNormalizer():
       >>> util.normalize(u'simpleandsafe', locale='de')
       'simpleandsafe'
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de')
       'text with umlaeut'
 
     Make sure we get the de normalizer as there's no special one for de-DE
     registered.
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de-DE')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de-DE')
       'text with umlaeut'
 
       >>> util.normalize(u'simpleandsafe', locale='pt-BR')
@@ -189,7 +195,7 @@ def testLocaleAwareFileNameNormalizer():
     Make sure we also handle POSIX-format locale identifiers,
     for backwards-compatibility with an earlier version of plone.i18n.
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de_DE')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de_DE')
       'text with umlaeut'
 
     """
@@ -216,7 +222,7 @@ def testURLNormalizer():
       >>> util.normalize(u"Doe, Joe")
       'doe-joe'
 
-      >>> util.normalize(six.text_type("umläut.doc", 'utf-8'))
+      >>> util.normalize(utf8unicode("umläut.doc"))
       'umlaut.doc'
 
       >>> util.normalize('quote "this"!')
@@ -231,7 +237,7 @@ def testURLNormalizer():
       >>> util.normalize("I'm a big file.TXT")
       'im-a-big-file.txt'
 
-      >>> util.normalize(six.text_type("rest `n` peace", 'utf-8'))
+      >>> util.normalize(u"rest `n` peace")
       'rest-n-peace'
 
       >>> len(util.normalize(u'aa' * 2000))
@@ -256,13 +262,13 @@ def testLocaleAwareURLNormalizer():
       >>> util.normalize(u'simpleandsafe', locale='de')
       'simpleandsafe'
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de')
       'text-with-umlaeut'
 
     Make sure we get the de normalizer as there's no special one for de-DE
     registered.
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de-DE')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de-DE')
       'text-with-umlaeut'
 
       >>> util.normalize(u'simpleandsafe', locale='pt-BR')
@@ -277,7 +283,7 @@ def testLocaleAwareURLNormalizer():
     Make sure we also handle POSIX-format locale identifiers,
     for backwards-compatibility with an earlier version of plone.i18n.
 
-      >>> util.normalize(six.text_type('text with umläut', 'utf-8'), locale='de_DE')
+      >>> util.normalize(utf8unicode('text with umläut'), locale='de_DE')
       'text-with-umlaeut'
 
     """